Restructure around real components
