+++ www/config/onCreateWebpackConfig.js
@@ -1,3 +1,5 @@
+// @ts-check
+
 const path = require("path")
 const WebpackNotifierPlugin = require("webpack-notifier")
 
@@ -12,10 +14,7 @@ module.exports = function onCreateWebpackConfig({ actions }) {
       }),
     ],
     resolve: {
-      alias: {
-        palette: path.resolve(process.cwd(), "src"),
-      },
-      modules: [path.resolve(process.cwd(), "src"), "node_modules"],
+      modules: [path.resolve(__dirname, "../src"), "node_modules"],
     },
   })
 }
