+++ www/gatsby-browser.js
@@ -1,3 +1,5 @@
+// @ts-check
+
 /**
  * Implement Gatsby's Browser APIs in this file.
  *
