+++ www/gatsby-node.js
@@ -1,11 +1,13 @@
+// @ts-check
+
 /**
  * Implement Gatsby's Node APIs in this file.
  *
  * See: https://www.gatsbyjs.org/docs/node-apis/
  */
 
-const onCreateNode = require("./system/onCreateNode")
-const onCreateWebpackConfig = require("./system/onCreateWebpackConfig")
+const onCreateNode = require("./config/onCreateNode")
+const onCreateWebpackConfig = require("./config/onCreateWebpackConfig")
 
 // Modifies the GraphQL schema
 exports.onCreateNode = onCreateNode
