+++ www/gatsby-ssr.js
@@ -1,3 +1,5 @@
+// @ts-check
+
 /**
  * Implement Gatsby's SSR (Server Side Rendering) APIs in this file.
  *
