+++ www/src/components/Header.tsx
@@ -1,10 +1,16 @@
 import { Link } from "gatsby"
 import React from "react"
 
-export const Header = ({ siteTitle }) => (
-  <div>
-    <h1>
-      <Link to="/">{siteTitle}</Link>
-    </h1>
-  </div>
-)
+interface HeaderProps {
+  title: string
+}
+
+export const Header: React.SFC<HeaderProps> = ({ title }) => {
+  return (
+    <div>
+      <h1>
+        <Link to="/">{title}</Link>
+      </h1>
+    </div>
+  )
+}
