+++ www/src/pages/components/Avatar.mdx
@@ -3,4 +3,10 @@ name: "avatar"
 foo: "bar"
 ---
 
-# HI!
+# HIheyy!
+
+hello there
+
+export const notes = {
+  hello: "world",
+}
